Stop linking nvim to vim config
Represents a shift in thinking. I had previously wanted to keep nvim and
vim as close as possible in terms of functionality, but as I've moved
more and more into the neovim ecosystem I've had to include more and
more `if has('nvim')` statements, causing more and more divergence. The
vim setup I have is sufficient for a non-IDE experience while still
providing a nice experience. The nvim setup is becoming more and more
like an IDE, and the config is becoming increasingly complex and
scattered, with keybindings sometimes specified in lua and sometimes in
vimscript. This commit will be the first step towards writing my neovim
config entirely in lua.