Remove git status binding override
Fugitive git status is better than telescope's as it provides explorable
diffs, staging with a single key, and commit authoring. Not to mention
it's a buffer, so all my usual keybindings work. Finding Telescope,
Mason, and Lazy's "modal" occasionally frustrating because they are not
buffers and only keep some keybindings.