Make find mappings consistent
While this makes the find commands slightly less convenient (eg the
change from `<leader>f` to `<leader>ff`, it makes the mappings far more
consistent and in line with the language of basic vim commands. For
example in the past I had to remember the specific key bindings for each
use-case, now I just have to remember a single key for the kind of
search I'm doing, because everything is being searched with a
"<Leader>f(ind)" prefix.