fix(nvim): Startup error caused by use-before-definition
Issue was that the call to `vim.lsp.config.eslint` would fail because
`vim.lsp.config` wasn't defined. It's only defined when the lspconfig
plugin is loaded, which was being loaded lazily thanks to the `Lazy`
plugin manager. By changing the setup of the LSP servers such that it
is part of the lazy package's `config` function, we ensure we load the
plugin before the call is made.

Incudes a refactor of the LSP configuration by having each become its
own lua module. This allows separation between the configuration of all
the LSPs from the individual ones, as previously it was difficult to
read from the large single table.