feat(alacritty): Add initial config
Note that this is only for v0.12, since that's what's available in the
current PopOS apt repos. Every other device I have is using a later
version and will need the toml equivalent since that was changed in
v0.13. For now I plan to define both. Hopefully in December I'll be able
to upgrade to the new PopOS release which will have a later version and
I can delete the yaml definition.